netcdf-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
xarray==2023.10.1
netcdf4==1.6.5
rasterio==1.3.9
pillow==10.1.0
matplotlib==3.8.2
contextily==1.4.0
pyproj==3.6.1